requirements.txt
# FinOps Analytics Platform — Dependencies

# ── Core data stack ───────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# ── Database ──────────────────────────────────────────────
# sqlite3 is built into Python — no install needed

# ── Web framework ─────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0              # FastAPI TestClient

# ── LangChain / LangGraph (1.x) ───────────────────────────
langchain>=1.0.0
langchain-core>=1.0.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0

# ── RAG ───────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# ── Visualisation ─────────────────────────────────────────
matplotlib>=3.8.0

# ── Testing ───────────────────────────────────────────────
pytest>=8.0.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv>=1.0.0